requirements.txt
openai>=1.40.0
jsonschema>=4.23.0
typer>=0.12.0
python-dotenv>=1.0.1
pytest>=8.3.0
markdownify>=0.13.0
